Split ICU4X data generation from `boa_icu_provider`
