+++ Cargo.toml
@@ -10,6 +10,9 @@ members = [
     # TESTS
     "tests/*",
 
+    # TOOLS
+    "tools/*",
+
     # OTHERS
     "examples",
     "cli",
@@ -89,7 +92,7 @@ icu_datagen = { version = "~1.4.1", default-features = false }
 icu_provider_adapters = { version = "~1.4.0", default-features = false }
 icu_provider_blob = { version = "~1.4.0", default-features = false }
 icu_properties = { version = "~1.4.0", default-features = true }
-icu_normalizer = { version = "~1.4.1", default-features = true }
+icu_normalizer = { version = "~1.4.1", default-features = false }
 icu_decimal = { version = "~1.4.0", default-features = false }
 writeable = "~0.5.4"
 yoke = "~0.7.3"
