+++ core/engine/src/builtins/array/mod.rs
@@ -2175,13 +2175,13 @@ impl Array {
 
         // 3. Let separator be the implementation-defined list-separator String value appropriate for the host environment's current locale (such as ", ").
         let separator = {
-            #[cfg(feature = "intl")]
+            #[cfg(feature = "intl_core")]
             {
                 // TODO: this should eventually return a locale-sensitive separator.
                 utf16!(", ")
             }
 
-            #[cfg(not(feature = "intl"))]
+            #[cfg(not(feature = "intl_core"))]
             {
                 utf16!(", ")
             }
