+++ core/engine/src/object/datatypes.rs
@@ -198,7 +198,7 @@ impl<T: ToOwned + ?Sized> JsData for Cow<'static, T> {}
 
 impl<T> JsData for Cell<Option<T>> {}
 
-#[cfg(feature = "intl")]
+#[cfg(feature = "intl_core")]
 default_impls!(icu_locid::Locale);
 
 impl<T: Trace + ?Sized> JsData for Gc<T> {}
