+++ core/icu_provider/README.md
@@ -1,12 +1,4 @@
 # boa_icu_provider
 
-`boa_icu_provider` generates and defines the [ICU4X](https://github.com/unicode-org/icu4x) data provider
+`boa_icu_provider` defines the [ICU4X](https://github.com/unicode-org/icu4x) data provider
 used in the Boa engine to enable internationalization functionality.
-
-## Datagen
-
-To regenerate the data:
-
-```bash
-$ cargo run --release --bin boa_datagen --features bin
-```
