+++ tools/gen-icu4x-data/README.md
@@ -1,0 +1,12 @@
+# boa_icu_provider
+
+`gen-icu4x-data` generates the [ICU4X](https://github.com/unicode-org/icu4x) data provider
+for `boa_icu_provider`.
+
+## Datagen
+
+To regenerate the data:
+
+```bash
+$ cargo run --release --bin gen-icu4x-data
+```
